+++ knp/base-framework/impl/visualizer/generate_files.h
